+++ CHANGES.md
@@ -4,6 +4,10 @@ moodle-tool_opcache
 Changes
 -------
 
+### Unreleased
+
+* 2017-08-14 - Add nagios check as CLI script
+
 ### Release v3.2-r1
 
 * 2017-08-08 - Initial version